requirements.txt
kodi.tools==1.1.0
Kodistubs==19.0.3
pytest==6.2.5
script.module.akl==1.1.0-rc0
requests==2.22.0